+++ app/custom-validators/helpers/utils/create-error-message.js
@@ -1,12 +1,12 @@
-function createErrorMessage(errorMessages, index, errorKey, fieldKeySuffix, focusSuffix) {
+function createErrorMessage(errorMessages, index, errorKey, focusSuffix) {
   const errorMessage = {
     inline: `${errorKey}.inline`,
     summary: `${errorKey}.summary`,
     variables: {
       indexKey: index + 1,
     },
-    fieldKeySuffix,
-    focusSuffix: [`[${focusSuffix}]`],
+    fieldKeySuffix: focusSuffix,
+    focusSuffix: [focusSuffix],
   };
 
   errorMessages.push(errorMessage);